+++ flutter_google_places_sdk/example/lib/constants.dart
@@ -1,0 +1,10 @@
+library constants;
+
+import 'package:flutter/material.dart';
+
+/// note: do NOT store your api key in here or in the code at all.
+/// use an external source such as file or firebase remote config
+const INITIAL_API_KEY = '';
+
+/// Initial value that is used for the locale
+const INITIAL_LOCALE = Locale('en');